+++ Engine/src/Architecture/TextureManager/TextureManager.h
@@ -10,16 +10,16 @@ namespace IonixEngine
 	class TextureManager
 	{
 	private:
-		std::map<uint64_t, TextureData> textureDict;
+		std::map<uint32_t, TextureData> textureDict;
 		TextureData errorTexture;
 		SDL_Renderer* renderer;
 	public:
 		TextureManager();
 		void AddTexture(std::string filepath,std::string alias);
 		TextureData& GetTexture(std::string alias);
-		TextureData& GetTexture(uint64_t hash);
+		TextureData& GetTexture(uint32_t hash);
 
-		SDL_Texture* GetRawTexture(uint64_t hash);
+		SDL_Texture* GetRawTexture(uint32_t hash);
 
 		void Init();
 		void Shutdown();
